tweeking reward parameters

+++ game/game.py
@@ -4,7 +4,7 @@ import sys
 
 class Game:
     def __init__(self):
-        self.board = np.zeros((4, 4), dtype=np.float)
+        self.board = np.zeros((4, 4), dtype=np.int)
         self.game_over = False
         self.new_board = np.zeros((4, 4), dtype=np.int)
         self.step = 0
